archive/simple_requirements.txt
Flask==2.3.3
google-genai==0.3.0
fpdf2==2.7.6
Werkzeug==2.3.7
gunicorn==21.2.0
